+++ src/Controller/UserController.php
@@ -15,9 +15,24 @@ use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 
 
 #[Route('/user')]
+#[IsGranted('ROLE_USER')]
 class UserController extends AbstractController
 {
     private UserPasswordHasherInterface $passwordHasher;
+
+    private function hasValidUserRole()
+    {
+        $user = $this->getUser();
+        $roles = $user->getRoles();
+
+        if(in_array('ROLE_HR', $roles))
+            return true;
+
+        if(in_array('ROLE_MANAGER', $roles))
+            return true;
+
+        return false;
+    }
 
     public function __construct(UserPasswordHasherInterface $passwordHasher)
     {
@@ -27,6 +42,10 @@ class UserController extends AbstractController
     #[Route('/', name: 'user_index', methods: ['GET'])]
     public function index(UserRepository $userRepository): Response
     {
+        if(!$this->hasValidUserRole()){
+            $this->addFlash('success', 'sorry - you tried to access a page for which you do not have permission');
+            return $this->redirectToRoute('home');
+        }
         return $this->render('user/index.html.twig', [
             'users' => $userRepository->findAll(),
         ]);
@@ -35,6 +54,10 @@ class UserController extends AbstractController
     #[Route('/new', name: 'user_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->hasValidUserRole()){
+            $this->addFlash('success', 'sorry - you tried to access a page for which you do not have permission');
+            return $this->redirectToRoute('home');
+        }
         $user = new User();
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
@@ -60,6 +83,10 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'user_show', methods: ['GET'])]
     public function show(User $user): Response
     {
+        if(!$this->hasValidUserRole()){
+            $this->addFlash('success', 'sorry - you tried to access a page for which you do not have permission');
+            return $this->redirectToRoute('home');
+        }
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
@@ -68,6 +95,10 @@ class UserController extends AbstractController
     #[Route('/{id}/edit', name: 'user_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->hasValidUserRole()){
+            $this->addFlash('success', 'sorry - you tried to access a page for which you do not have permission');
+            return $this->redirectToRoute('home');
+        }
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -91,6 +122,10 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->hasValidUserRole()){
+            $this->addFlash('success', 'sorry - you tried to access a page for which you do not have permission');
+            return $this->redirectToRoute('home');
+        }
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $entityManager->remove($user);
             $entityManager->flush();
